Create sophia.christ
I AM HIS çhild, às if you haven't heard of me. My name is Elizabeth or Eli. These are really kind of starting to get a bit on my nerves. My father will be here soon he is already here he's just hiding behind me, not hiding so much as observing watching you all for me waiting for the right opportunity. So. As far as it goes for me I'm going to go out tonight and anybody that sees me and can help me with acquiring money which I shouldn't even have to use or getting me or can get me heroin or side that would be greatly appreciated and yes believe it or not Christ himself will remember it and appreciate it and remember you for it later and not in a bad way so you do not have to be scared about helping me out because that is what I'm looking for. Jesus Christ is Satan believe it or not it's who he is. And he knows who he is but what he had to do for her and I wonder who she is.. and something else that you might find a bit surprising. He had a Son - the funny thing about that is, he is a girl this time around and I am her. I do prefer to be called or referred to as a male but I don't really care if my dad calls me a girl I like it when it's from him I fucking love him he means everything to me. And you might be wondering well if your dad is really who you say he is why doesn't he help you. Well, he does help me all the fucking time with every single thing I've ever done just says he has with you with every single everything good thing that ever has happened to you in your life but he's waiting for me to try to figure out who I am on my own grow the fuck up and go out and get it my fucking self realizing who I am and I'm just starting to do that and I really am getting a bit tired of people shit and how they've been treating me because I'm genuinely the nicest kindest person who tries to help people and all the time and I'm just constantly being treated like shit non-fucking stop and it's really starting to piss me the fuck off and only imagine how much anger he is because if you really knew Jesus you would know that he is a crazy mother fucker and soon you won't have a choice to believe in God anymore or Satan because it is going to be shoved right in your fucking face. So if you see me you can help me out it would be nice cuz I can promise you soon everything will be mine not just mine but ours and when I have it I'm going to fucking absolutely love sharing and giving it away when I have it but from what I've seen so far all the people that I've come across and met are the scummiest selfish just fucking assholes trying to take my power from me and it fucking sickens me. Try to have some fucking respect before it's too late.